pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "aivonx"
version = "0.13.5"
requires-python = ">=3.12, <3.13"
description = "aivonx"

authors = [
    { name = "Xuan‑You Lin", email = "a0985821880@gmail.com" }
]

dependencies = [
    "python-dotenv>=1.2.1",
    "gunicorn>=23.0.0",
    "uvicorn>=0.38.0",
    "django>=5.2.9",
    "django-asgi-lifespan>=0.5.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "httpx>=0.28.1",
    "apscheduler>=3.11.1",
    "python-json-logger>=4.0.0",
    "whitenoise>=6.11.0",
    "django-redis>=6.0.0",
    "redis>=7.1.0",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "ipykernel",
    "ipywidgets",
    "nest_asyncio",
    "jupyterlab_widgets",
]
web = [
    "mkdocs",
    "mkdocs-material",
    "mkdocs-static-i18n",
    "mkdocstrings[python]",
]
safe = [
    "bandit",
    "safety",
]

[tool.hatch.build.targets.wheel]
packages = ["src/"]
